requirements.txt
zeep
python-dotenv
requests
mcp
uvicorn
bs4